requirements.txt
fastapi
pydantic
Pillow
requests
cairosvg
python-multipart
uvicorn